Updated MDSystemVersion stuff for MDMountainLion.
Fixed warnings on string formats by casting parameters.

git-svn-id: file:///Users/mdouma46/Developer/sourceFinagler/branches/2.0.3@61 c3d3028b-b5cd-4c35-8691-6fb8b2347b27

+++ MDAppKitAdditions.m
@@ -88,7 +88,7 @@ NSString *NSStringFromDefaultsKeyPath(NSString *defaultsKey) {
 		NSInteger greenInt = green * 255.99999f;
 		NSInteger blueInt = blue * 255.99999f;
 		
-		hexValue = [NSString stringWithFormat:@"#%02x%02x%02x", redInt, greenInt, blueInt];
+		hexValue = [NSString stringWithFormat:@"#%02x%02x%02x", (unsigned int)redInt, (unsigned int)greenInt, (unsigned int)blueInt];
 		
 	}
