+++ MDFoundationAdditions.m
@@ -212,13 +212,13 @@ SInt32 MDGetSystemVersion() {
 				NSString *prefix = [filename substringToIndex:29];
 				NSString *lastCharacter = [filename substringFromIndex:[filename length] - 1];
 				
-				filename = [NSString stringWithFormat:@"%@%C%@", prefix, 0x2026, lastCharacter];				
+				filename = [NSString stringWithFormat:@"%@%C%@", prefix, (unsigned short)0x2026, lastCharacter];				
 				
 			} else if ([components count] > 1) {
 				NSUInteger lastComponentLength = [[components lastObject] length];
 				
 				NSString *suffix = [filename substringFromIndex:[filename length] - lastComponentLength - 1];
-				suffix = [[NSString stringWithFormat:@"%C", 0x2026] stringByAppendingString:suffix];
+				suffix = [[NSString stringWithFormat:@"%C", (unsigned short)0x2026] stringByAppendingString:suffix];
 				
 				NSString *prefix = [filename substringToIndex:[filename length] - lastComponentLength - 2];
 				
@@ -234,7 +234,7 @@ SInt32 MDGetSystemVersion() {
 			NSString *suffix = [filename substringFromIndex:(pRange.location - 1)];
 			NSString *prefix = [filename substringToIndex:(pRange.location - 1)];
 			
-			suffix = [[NSString stringWithFormat:@"%C", 0x2026] stringByAppendingString:suffix];
+			suffix = [[NSString stringWithFormat:@"%C", (unsigned short)0x2026] stringByAppendingString:suffix];
 			
 			NSUInteger allowedPrefixLength = (31 - [suffix length]);
